+++ core/CCArray.mli
@@ -70,6 +70,15 @@ module type S = sig
   (** [find f a] returns [Some y] if there is an element [x] such
       that [f x = Some y], else it returns [None] *)
 
+  val findi : (int -> 'a -> 'b option) -> 'a t -> 'b option
+  (** Like {!find}, but also pass the index to the predicate function.
+      @since 0.3.4 *)
+
+  val find_idx : ('a -> bool) -> 'a t -> (int * 'a) option
+  (** [find p x] returns [Some (i,x)] where [x] is the [i]-th element of [l],
+      and [p x] holds. Otherwise returns [None]
+      @since 0.3.4 *)
+
   val lookup : ?cmp:'a ord -> 'a -> 'a t -> int option
   (** Lookup the index of some value in a sorted array.
       @return [None] if the key is not present, or