+++ core/CCInt.ml
@@ -26,9 +26,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 type t = int
 
-let equal a b = a=b
+let equal (a:int) b = a=b
 
-let compare a b = Pervasives.compare a b
+let compare (a:int) b = Pervasives.compare a b
 
 let hash i = i land max_int
 
@@ -46,4 +46,4 @@ let random_small = random 100
 let random_range i j st = i + random (j-i) st
 
 let pp buf = Printf.bprintf buf "%d"
-let print fmt = Format.fprintf fmt "%d"
+let print fmt = Format.pp_print_int fmt